only use new inner on Linux
